Also fix the spaces chats screen by using the same feature as the overview widget
