+++ app/lib/features/space/widgets/space_sections/chats_section.dart
@@ -1,11 +1,7 @@
-import 'package:acter/common/providers/chat_providers.dart';
 import 'package:acter/common/providers/space_providers.dart';
 import 'package:acter/common/utils/routes.dart';
-import 'package:acter/common/widgets/chat/convo_card.dart';
-import 'package:acter/common/widgets/chat/convo_hierarchy_card.dart';
-import 'package:acter/common/widgets/chat/loading_convo_card.dart';
+import 'package:acter/features/space/widgets/related/chats_helpers.dart';
 import 'package:acter/features/space/widgets/space_sections/section_header.dart';
-import 'package:acter/router/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
@@ -85,65 +81,8 @@ class ChatsSection extends ConsumerWidget {
           ),
         ),
         chatsListUI(ref, chats, chatsLimit),
-        if (renderRemote) renderFurther(context, ref, moreCount),
+        if (renderRemote) renderFurther(context, ref, spaceId, moreCount),
       ],
-    );
-  }
-
-  Widget chatsListUI(WidgetRef ref, List<String> chats, int chatsLimit) {
-    return ListView.builder(
-      shrinkWrap: true,
-      itemCount: chatsLimit,
-      padding: EdgeInsets.zero,
-      physics: const NeverScrollableScrollPhysics(),
-      itemBuilder: (context, index) {
-        final roomId = chats[index];
-        return ref.watch(chatProvider(roomId)).when(
-              data: (room) => ConvoCard(
-                room: room,
-                showParents: false,
-                showSelectedIndication: false,
-                onTap: () => goToChat(context, roomId),
-              ),
-              error: (error, stack) => ListTile(
-                title: Text(roomId),
-                subtitle: Text(L10n.of(context).loadingFailed(error)),
-              ),
-              loading: () => LoadingConvoCard(roomId: roomId),
-            );
-      },
-    );
-  }
-
-  Widget renderFurther(BuildContext context, WidgetRef ref, int maxItems) {
-    final remoteChats = ref.watch(remoteChatRelationsProvider(spaceId));
-
-    return remoteChats.when(
-      data: (chats) {
-        if (chats.isEmpty) {
-          return const SizedBox.shrink();
-        }
-
-        return ListView.builder(
-          shrinkWrap: true,
-          padding: EdgeInsets.zero,
-          physics: const NeverScrollableScrollPhysics(),
-          itemCount: maxItems,
-          itemBuilder: (context, idx) {
-            final item = chats[idx];
-            return ConvoHierarchyCard(
-              showIconIfSuggested: true,
-              parentId: spaceId,
-              roomInfo: item,
-            );
-          },
-        );
-      },
-      error: (e, s) =>
-          Card(child: Text(L10n.of(context).errorLoadingRelatedChats(e))),
-      loading: () => Skeletonizer(
-        child: Card(child: Text(L10n.of(context).loadingOtherChats)),
-      ),
     );
   }
 }
